refactor: 챗 초대 사이드바 리팩터
