+++ src/pages/chat/components/sidebar/chat-member-tap.tsx
@@ -1,5 +1,5 @@
 import ChatMemberInquireTap from 'pages/chat/components/sidebar/inquire-tap/chat-member-inquire-tap';
-import ChatMemberInviteTap from 'pages/chat/components/sidebar/chat-member-invite-tap';
+import ChatMemberInviteTap from 'pages/chat/components/sidebar/invite-tap/chat-member-invite-tap';
 import { useState } from 'react';
 
 const ChatMemberTap = () => {
